refactor: Revert moving `types.ts` and `database.types.ts`
