+++ src/app/login/_components/RegisterModal.tsx
@@ -4,7 +4,7 @@ import { FilledButton } from "@/elements/FilledButton";
 import { LabelledInput } from "@/elements/LabelledInput";
 import { Modal } from "@/elements/Modal";
 import { Notification } from "@/elements/Notification";
-import type { ActionState } from "@/types/app.types";
+import type { ActionState } from "@/lib/types";
 
 type RegisterModalProps = {
   isOpen: boolean;
