+++ src/app/tournament/[uuid]/_components/DeletePlayerModal.tsx
@@ -5,7 +5,7 @@ import { Notification } from "@/elements/Notification";
 import { RoundedListbox } from "@/elements/RoundedListbox";
 import { usePlayerMutations } from "@/hooks/players/usePlayerMutations";
 import { usePlayers } from "@/hooks/players/usePlayers";
-import type { Player } from "@/types/app.types";
+import type { Player } from "@/lib/types";
 
 type DeletePlayerModalProps = {
   isOpen: boolean;
