+++ src/app/tournament/[uuid]/_components/TableList.tsx
@@ -3,7 +3,7 @@ import { twMerge } from "tailwind-merge";
 import { FilledButton } from "@/elements/FilledButton";
 import { useSessions } from "@/hooks/sessions/useSessions";
 import { useTableMutations } from "@/hooks/tables/useTableMutations";
-import type { Table } from "@/types/app.types";
+import type { Table } from "@/lib/types";
 import { TableSeat } from "./TableSeat";
 
 type TableListProps = {
