+++ src/app/tournament/[uuid]/analytics/_components/PlayerAnalytics.tsx
@@ -2,7 +2,7 @@ import { LineGraph } from "@/elements/LineGraph";
 import { useLogs } from "@/hooks/logs/useLogs";
 import { useTournaments } from "@/hooks/tournaments/useTournaments";
 import { getPointHistory } from "@/lib/scores";
-import type { Player } from "@/types/app.types";
+import type { Player } from "@/lib/types";
 
 type PlayerAnalyticsProps = {
   player: Player;
