+++ src/app/tournament/[uuid]/logs/page.tsx
@@ -6,8 +6,8 @@ import { IconButton } from "@/elements/IconButton";
 import { useLogs } from "@/hooks/logs/useLogs";
 import { usePlayers } from "@/hooks/players/usePlayers";
 import { useSessions } from "@/hooks/sessions/useSessions";
+import type { Log, LogSearchTag } from "@/lib/types";
 import { normalizeText } from "@/lib/utils";
-import type { Log, LogSearchTag } from "@/types/app.types";
 import { LogList } from "./_components/LogList";
 import { LogSearchBar } from "./_components/LogSearchBar";
