+++ src/app/tournament/[uuid]/sessions/_components/ViewGraphModal.tsx
@@ -2,7 +2,7 @@ import { useMemo } from "react";
 import { BarChart } from "@/elements/BarChart";
 import { Modal } from "@/elements/Modal";
 import { rankPlayers } from "@/lib/scores";
-import type { Player } from "@/types/app.types";
+import type { Player } from "@/lib/types";
 
 type ViewGraphModalProps = {
   players: Player[];
